% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{cellmarker}
\alias{cellmarker}
\title{Title}
\usage{
cellmarker(
  marker_list,
  outdir,
  prefix = "cellmarker",
  species = "Mouse",
  showCategory = 20,
  sep = "\\t",
 
    databese = "/annoroad/data1/bioinfo/PROJECT/RD/Cooperation/RNA/sRNA/ngs_bioinfo/TET-1487/yaomengcheng/Analysis-test/clusterProfiler/database",
  w = 12,
  h = 12
)
}
\arguments{
\item{marker_list}{marker list from seurat diff output}

\item{outdir}{outdir}

\item{prefix}{outfile prefix}

\item{species}{species for analysis, must be in single,Mouse,Human,all}

\item{showCategory}{show top numbers for output pdf file}

\item{databese}{cellmarker annotation file dir}

\item{w}{The width of output pdf}

\item{h}{The height of output pdf}
}
\value{
The compareCluster result of output
}
\description{
Title
}
\examples{
            gene_list<-seurat_marker(markerfile,sep=",",colname="cluster")
            cellmarker(gene_list,outdir,species="Mouse",prefix="cell_marker")
}
