% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnetPlot.R
\name{cneteplot}
\alias{cneteplot}
\title{cneteplot}
\usage{
cneteplot(data, showCategory, scale, layout, colors)
}
\arguments{
\item{data}{result from enrichKEGG}

\item{showCategory}{number of enriched terms to display(default 5)}

\item{scale}{size of nodes(default 2)}

\item{layout}{"force" or "circular"(default force)}

\item{colors}{a color list(default rainbow(5))}
}
\value{
a echart plot
}
\description{
plot cnet of enrichKEGG
}
\examples{
cneteplot(enrichKK)
}
