% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compare}
\alias{compare}
\title{Title}
\usage{
compare(
  markerfile,
  outdir,
  genefile,
  prefix = "compare_cluster",
  species = "Mus_musculus.GRCm38.90.chr",
  indir = "/annoroad/bioinfo/PMO/database/config_database",
  python = "/annoroad/data1/bioinfo/PMO/yaomengcheng/Anaconda3/bin/python",
 
    script = "/annoroad/data1/bioinfo/PROJECT/RD/Cooperation/RNA/sRNA/ngs_bioinfo/TET-1487/yaomengcheng/Analysis-test/clusterProfiler/gene.go.py",
  go.class = "/annoroad/bioinfo/PMO/database/GO/current/go.class",
  sep = ",",
  colname = "cluster",
  pvalueCutoff = 0.1,
  pAdjustMethod = "BH",
  qvalueCutoff = 1
)
}
\arguments{
\item{markerfile}{markerfile from seurat diff output}

\item{outdir}{output dir for output file}

\item{genefile}{ensemble and symbol file from cellrange output}

\item{prefix}{output's prefix}

\item{species}{gene go annotation for species,must in list for annorad annotation list}

\item{indir}{database indir}

\item{python}{which python for dealing with go annodataion file}

\item{script}{which script for dealing with go annodataion file}

\item{go.class}{Go Term and Go Term namse file}

\item{sep}{marker file sep}

\item{colname}{marker for group colnames}

\item{pvalueCutoff}{compareCluster analysis for param,pvalueCutoff}

\item{pAdjustMethod}{compareCluster analysis for param,pAdjustMethod}

\item{qvalueCutoff}{compareCluster analysis for param,qvalueCutoff}
}
\value{
S4 object for result, from clusterprofiler
}
\description{
Title
}
\examples{
              x<-compare(markerfile,outdir,genefile,prefix=prefix)
}
