% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{compare_plot}
\alias{compare_plot}
\title{Title}
\usage{
compare_plot(x, outdir, prefix, w = 12, h = 8, showCategory = 10)
}
\arguments{
\item{x}{S4 object for result  from clusterprofiler}

\item{outdir}{output dir}

\item{prefix}{outputfile prefix}

\item{w}{the width for pdf file}

\item{h}{the heigh for pdf file}

\item{showCategory}{show top numbers of analysis pdf file}
}
\value{

}
\description{
Title
}
\examples{
             compare_plot(x,outdir,prefix,w=12,h=8,showCategory=20)
}
