% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichKK}
\alias{enrichKK}
\title{result of enrichKEGG}
\format{
An object of class \code{enrichResult} with 15 rows and 9 columns.
}
\usage{
enrichKK
}
\description{
A dataset containing the results .
}
\keyword{datasets}
