% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ensemble2symbol}
\alias{ensemble2symbol}
\title{Title get gene symbol from ensemble name}
\usage{
ensemble2symbol(genefile, term2gene, sep = "\\t")
}
\arguments{
\item{genefile}{genes.tsv from cellranger output}

\item{term2gene}{data.fram, the first col is Go term names and the second col is Gene name}

\item{sep}{genefile sep,default sep="\t"}
}
\value{
term2gene  data.fram, the first col is Go term names and the second col is Gene name(symbol)
}
\description{
Title get gene symbol from ensemble name
}
\examples{
term2gene<-ensemble2symbol(genefile,term2gene)
}
