% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gene_ko_update}
\alias{gene_ko_update}
\title{Title gene_ko_update(not use config.ini annotation file, using result of blastx and updata by python script with the same of annorad methods)}
\usage{
gene_ko_update(
  outdir,
  go_gile,
  python_YMC = "/annoroad/data1/bioinfo/PMO/yaomengcheng/Anaconda3/bin/python",
 
    anno = "/annoroad/data1/bioinfo/PMO/Public/Pipeline/Stable/RNA/Public_Module/current/KEGG//anno.py",
  perl = "/annoroad/share/software/install//perl-5.16.2/bin/perl",
 
    extract_ko = "/annoroad/data1/bioinfo/PMO/Public/Pipeline/Stable/RNA/Public_Module/current/KEGG/extract_ko_from_bgi_result.pl",
 
    sprot.anno = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/Uniport/current/data/sprot.anno"
)
}
\arguments{
\item{outdir}{which mkdir tmp dirs for output}

\item{go_gile}{which come from  config.ini annotation file, but not using this annotation file and provided dirnames for blastx results}

\item{python_YMC}{which python to use}

\item{anno}{which python script anno.py to use}

\item{perl}{which perl to use}

\item{extract_ko}{which extract_ko script to use}

\item{sprot.anno}{which sprot.anno file to use}
}
\value{
ko_updata.list a data.frame(the first colum is Gene and the second colum is ko)
}
\description{
Title gene_ko_update(not use config.ini annotation file, using result of blastx and updata by python script with the same of annorad methods)
}
\examples{
gene_ko_update(outdir,ko_gile)
}
