% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{go_enrich_local_compareCluster}
\alias{go_enrich_local_compareCluster}
\title{Title kegg_enrich_local_compareCluster}
\usage{
go_enrich_local_compareCluster(
  species,
  gene_list,
  outdir,
  indir = "/annoroad/bioinfo/PMO/database/config_database",
  go.class = "/annoroad/bioinfo/PMO/database/GO/current/go.class",
  ensemble_symbol_file = NULL,
  ont = "all",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.05
)
}
\arguments{
\item{species}{gene go annotation for species,must in list for annorad annotation list}

\item{gene_list}{a gene_list for enrichment, the gene list of length range from 1-max}

\item{outdir}{output dir for output file}

\item{indir}{database indir}

\item{go.class}{go.class file for go annotation}

\item{ensemble_symbol_file}{ensemble and symbol file from cellrange output}

\item{ont}{Ontology for go ananlysis[BP CC MF]}

\item{pvalueCutoff}{compareCluster analysis for param,pvalueCutoff}

\item{pAdjustMethod}{compareCluster analysis for param,pAdjustMethod}

\item{qvalueCutoff}{compareCluster analysis for param,qvalueCutoff}
}
\value{
S4 object for result, from clusterprofiler
}
\description{
Title kegg_enrich_local_compareCluster
}
\examples{
  x<-kegg_enrich_local_compareCluster(species,gene_list[[1]],outdir)
}
