% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{go_gene}
\alias{go_gene}
\title{Title go_gene}
\usage{
go_gene(
  infile,
  outfile,
  python = "/annoroad/data1/bioinfo/PMO/yaomengcheng/Anaconda3/bin/python",
 
    script = "/annoroad/data1/bioinfo/PROJECT/RD/Cooperation/RNA/sRNA/ngs_bioinfo/TET-1487/yaomengcheng/Analysis-test/clusterProfiler/gene.go.py"
)
}
\arguments{
\item{infile}{go annotation file name}

\item{outfile}{Gene GO txt file,just only two col, and the first col is gene(ensemble), the second col is GO term name}

\item{python}{absulately python path to deal go annotation file}

\item{script}{script to deal go annotation file}
}
\value{
data.fram, the first col is Go term names and the second col is Gene name
}
\description{
Title go_gene
}
\examples{
term2gene<-go_gene(infile,paste(outdir,"term2gene.xls",sep="/"),script=script,python=python)[,c("GO","Gene")]
}
