% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{go_Pathways}
\alias{go_Pathways}
\title{Title go.list or ko.list to list}
\usage{
go_Pathways(gmt.file)
}
\arguments{
\item{gmt.file}{}
}
\value{
list,
}
\description{
Title go.list or ko.list to list
}
\examples{
 ko2map<-go_Pathways('/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/KEGG/current/data/ko2map/ko2map.xls')
}
