% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{go_updata}
\alias{go_updata}
\title{Title go_updata(not use config.ini annotation file, using result of blastx and updata by python script with the same of annorad methods)}
\usage{
go_updata(
  outdir,
  go_gile,
  python_YMC = "/annoroad/data1/bioinfo/PMO/yaomengcheng/Anaconda3/bin/python",
 
    anno = "/annoroad/data1/bioinfo/PMO/Public/Pipeline/Stable/RNA/Public_Module/current/Function_GO//anno.py",
 
    update_go = "/annoroad/data1/bioinfo/PMO/Public/Pipeline/Stable/RNA/Public_Module/current/Function_GO///update_go.py",
 
    sprot.anno = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/Uniport/current/data/sprot.anno",
 
    go.class = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/GO/current/data/go.class",
 
    go.update = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/GO/current/data/go.update"
)
}
\arguments{
\item{outdir}{which mkdir tmp dirs for output}

\item{go_gile}{which come from  config.ini annotation file, but not using this annotation file and provided dirnames for blastx results}

\item{python_YMC}{which python to use}

\item{anno}{which python script anno.py to use}

\item{update_go}{which update_go script to use}

\item{sprot.anno}{which sprot.anno file to use}

\item{go.class}{which go.class file to use}

\item{go.update}{which go.update file to use}
}
\value{
go_updata.list  a file go_updata.list
}
\description{
Title go_updata(not use config.ini annotation file, using result of blastx and updata by python script with the same of annorad methods)
}
\examples{
go_updata(outdir,go_gile)
}
