% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{id2symbol}
\alias{id2symbol}
\title{Title id2symbol(just for apply funciton)}
\usage{
id2symbol(x, genes)
}
\arguments{
\item{X}{A vector of gene list}

\item{gene_id}{and gene_symbol data.frame}
}
\value{
a gene list with Gene_symbol vector
}
\description{
Title id2symbol(just for apply funciton)
}
\examples{
id2symbol(gene_list)
}
