% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{kegg_enrich_local_compareCluster}
\alias{kegg_enrich_local_compareCluster}
\title{Title kegg_enrich_local_compareCluster}
\usage{
kegg_enrich_local_compareCluster(
  species,
  gene_list,
  outdir,
  ko_color = "red",
  ensemble_symbol_file = NULL,
  type = "KEGG_annotate",
  indir = "/annoroad/bioinfo/PMO/database/config_database",
  category = "animal",
 
    term2name_pre = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/KEGG/current/data/pathway_",
 
    ko2map = "/annoroad/data1/bioinfo/PMO/Public/database/Public/Quarter/KEGG/current/data/ko2map/ko2map.xls",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.05
)
}
\arguments{
\item{species}{gene go annotation for species,must in list for annorad annotation list}

\item{gene_list}{a gene_list for enrichment, the gene list of length range from 1-max}

\item{outdir}{output dir for output file}

\item{ko_color}{output geneKO with color type,default red}

\item{ensemble_symbol_file}{ensemble and symbol file from cellrange output}

\item{indir}{database indir}

\item{category}{category for species[animal or plant... ]}

\item{term2name_pre}{term2name for kegg}

\item{ko2map}{ko2map file for ko and mapID}

\item{pvalueCutoff}{compareCluster analysis for param,pvalueCutoff}

\item{pAdjustMethod}{compareCluster analysis for param,pAdjustMethod}

\item{qvalueCutoff}{compareCluster analysis for param,qvalueCutoff}

\item{sep}{marker file sep}
}
\value{
S4 object for result, from clusterprofiler
}
\description{
Title kegg_enrich_local_compareCluster
}
\examples{
  x<-kegg_enrich_local_compareCluster(species,gene_list[[1]],outdir)
}
