% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mkdirs}
\alias{mkdirs}
\title{Title mkdirs funtion(if dir exist, do not del the dir)}
\usage{
mkdirs(outdir, fp)
}
\arguments{
\item{outdir}{outdir for mkdirs}

\item{fp}{dirnames for mkdirs}
}
\value{
mkdirs and retunr none
}
\description{
Title mkdirs funtion(if dir exist, do not del the dir)
}
\examples{
 mkdirs(outdir,'tmp/')
}
