% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{seurat_marker}
\alias{seurat_marker}
\title{Title seurat_marker}
\usage{
seurat_marker(markerfile, sep = ",", colname = "cluster")
}
\arguments{
\item{markerfile}{a markerfile from seurat diff for cluster}

\item{sep}{sep for marker file ,default=","}

\item{colname}{which col for group,defaul="cluster"}
}
\value{
list,
}
\description{
Title seurat_marker
}
\examples{
 seurat_marker(markerfile,sep=",",colname="cluster")
}
