% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{go_listfile}
\alias{go_listfile}
\title{Title go_listfile}
\usage{
go_listfile(species, indir = "/annoroad/bioinfo/PMO/database/config_database")
}
\arguments{
\item{species}{the ref gene annotation from annorad loacal database(Go Term)}

\item{indir}{the database indir}
}
\value{
go annotation file name
}
\description{
Title go_listfile
}
\examples{
go_list<-go_listfile("Mus_musculus.GRCm38.90.chr",indir='/annoroad/bioinfo/PMO/database/config_database')
}
